dev-7020 reverted configs back to original state, put Accordions back to use ui/componet
