+++ packages/core/helpers/coveUpdateWorker.ts
@@ -2,17 +2,14 @@
 // process those config entries and format old values into new
 import update_4_24_4 from './ver/4.23.4'
 import update_4_24_3 from './ver/4.24.3'
-import update_4_24_4 from './ver/4.24.4'
+import update_4_24_5 from './ver/4.24.5'
 
 export const coveUpdateWorker = config => {
   let genConfig = config
 
-  // v4.24.3
   genConfig = update_4_24_3(genConfig)
   genConfig = update_4_24_4(genConfig)
-
-  // v4.24.4
-  genConfig = update_4_24_4(genConfig)
+  genConfig = update_4_24_5(genConfig)
 
   return genConfig
 }
