+++ packages/markup-include/src/_stories/_mock/image-with-text.json
@@ -1,12 +1,9 @@
 {
-  "contentEditor": {
-    "inlineHTML": "<div>{{state}} does have a rate 130 compared to the over all rate of {{overall-rate}}</div>",
-    "markupVariables": [],
-    "showHeader": false,
-    "srcUrl": "/wcms/4.0/cdc-wp/data-presentation/examples/SSI-Image-With-Text.html",
-    "title": "<strong>Markup Include</strong> - Image with Text",
-    "useInlineHTML": true
-  },
+  "title": "<strong>Markup Include</strong> - Image with Text",
+  "showHeader": false,
+  "type": "markup-include",
+  "srcUrl": "/wcms/4.0/cdc-wp/data-presentation/examples/SSI-Image-With-Text.html",
+  "theme": "theme-slate",
   "data": [
     {
       "Race": "Hispanic or Latino",
@@ -29,16 +26,6 @@
       "Age-adjusted rate": "183.6"
     }
   ],
-  "legend": {},
-  "newViz": true,
-  "theme": "theme-slate",
-  "type": "markup-include",
-  "runtime": null,
-  "visual": {
-    "border": false,
-    "accent": true,
-    "background": true,
-    "hideBackgroundColor": false,
-    "borderColorTheme": false
-  }
+  "dataFileName": "valid-data-chart.csv",
+  "dataFileSourceType": "file"
 }
